requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bitarray==2.7.3
canonicaljson==2.0.0
certifi==2023.5.7
charset-normalizer==3.1.0
cytoolz==0.12.1
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
farcaster==0.7.7
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
multidict==6.0.4
parsimonious==0.9.0
pycryptodome==3.18.0
pydantic==1.10.8
pyhumps==3.8.0
python-decouple==3.8
python-dotenv==0.21.1
regex==2023.5.5
requests==2.31.0
rlp==3.0.0
soupsieve==2.4.1
toolz==0.12.0
tqdm==4.65.0
typing_extensions==4.6.2
urllib3==2.0.2
yarl==1.9.2
